auth_app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
aiosmtplib==3.0.1
jinja2==3.1.4
python-dotenv==1.0.1
# ── New: traffic, caching, rate-limiting ──
slowapi==0.1.9
redis[hiredis]==5.1.1
